Adding a method of doing a simple login and starting session.
Yuck.  Spring does not give you an easy way of returning * from
insert/update/delete and mapping it to an object.  The default row
mapping requires public methods to do that.

Hacking the method of providing an array of task ids.  Spring boot jdbc
does not natively interpret arrays and would require individual
statements in that case.

Starting to add Req/Resp (DTO) objects for transport.  Attempting to
figure out how to fix the save() method.

Fixing the package naming and class organization a bit.